test: split deploy safety and identity browser tests
