+++ docs/audits/openai-codex_data_Audit_Findings.md
@@ -222,7 +222,7 @@ Generated, vendored, ignored, or excluded areas:
 - **Evidence level:** E2
 - **Status:** Likely
 - **Rule or parameter:** Production schema changes should fail cleanly or use bounded locking under concurrent public traffic.
-- **Location:** `scripts/migrate.ts:511-531`, `db/migrations/20260814_public_pagination.sql:3`, `db/migrations/20260814_agreement_accession.sql:5`, `db/migrations/20260815_open_to_use.sql:1`, `db/migrations/20260814_world_root_topology.sql:6`, `test/deploy-safety.test.ts:534`, `test/deploy-safety.test.ts:591`, `test/deploy-safety.test.ts:646`
+- **Location:** `scripts/migrate.ts:511-531`, `db/migrations/20260814_public_pagination.sql:3`, `db/migrations/20260814_agreement_accession.sql:5`, `db/migrations/20260815_open_to_use.sql:1`, `db/migrations/20260814_world_root_topology.sql:6`, `test/deploy-safety-tests/production-and-local-targets.test.ts:143`, `test/deploy-safety-tests/production-and-local-targets.test.ts:195`, `test/deploy-safety-tests/migration-selection.test.ts:36`
 - **User or business harm:** Running these production migrations during traffic can hold live-table DDL locks long enough to stall agent writes, creating user-visible outages or release aborts at unpredictable points.
 - **Evidence:** `applyMigration()` wraps any migration lacking explicit `BEGIN/COMMIT` in one `sql.transaction(...)`. The `public-pagination` migration creates 10 indexes on busy live tables; `agreement-accession` alters `agreement_parties` and creates indexes/triggers; `open-to-use` alters `things`. None of those files set `lock_timeout`, `statement_timeout`, or use concurrent index creation. By contrast, the destructive `world-root-topology` migration explicitly sets short lock and statement timeouts before touching live topology.
 - **Safe reproduction:** Deterministic static proof from the cited files. Supporting test coverage is mostly shape/selection checks: the checked tests assert additive text and drift, but only `world-root-postgres` and `oauth-postgres` execute real upgraded-database migrations in Postgres.
