+++ docs/audits/whole-site_unshittify_2026-08-15.md
@@ -166,7 +166,7 @@ exists. This pattern is unsealed and carries no integrity claim.
 | --- | --- | --- | --- | --- | --- | --- |
 | EVD-001 | 16:04:21 | repo root | `audit_guard.py start --root ... --exclude "*.tsbuildinfo"` | 0 | `baseline_captured`, 834 files, 0 unreadable | Wrote manifest to OS temp only |
 | EVD-002 | ~16:12 | repo root | `npm run typecheck` (`tsc --noEmit`) | 0 | No type errors | None |
-| EVD-003 | ~16:14 | repo root | `npm test` | **1** | 313 tests, 310 pass, **3 fail**: `help-text` LLMS sync, and two `deploy-safety` tests | None (fixtures write to OS temp only, verified at `test/deploy-safety.test.ts:176,242,285`) |
+| EVD-003 | ~16:14 | repo root | `npm test` | **1** | 313 tests, 310 pass, **3 fail**: `help-text` LLMS sync, and two `deploy-safety` tests | None (fixtures write to OS temp only, verified at `test/deploy-safety-tests/preview-targets.test.ts:6`, `test/deploy-safety-tests/preview-targets.test.ts:57`, `test/helpers/deploy-safety-fixtures/preparation-worktree.ts:40`) |
 | EVD-004 | ~16:16 | skill dir | `git ls-tree -r HEAD` (hardened, `-C` repo) | 0 | 99 tracked paths at `1303a14` | None |
 | EVD-005 | ~16:17 | skill dir | `git hash-object` each tracked file vs `HEAD` | 0 | 95 identical, **4 differ** (`.gitignore`, `scripts/deploy.sh`, `src/door.ts`, `test/world-root.test.ts`), 0 missing | None |
 | EVD-006 | ~16:15 | skill dir | `git worktree list`, `git config --local --list` | 0 | Root is **bare**; 20 worktrees, 7 nested inside the root | None |
@@ -399,7 +399,7 @@ present there. Locations are repo-relative.
 - **Connections and similar locations checked:** `20260813_hosted_chat_signin.sql` also lacks the preamble, but it is additive table creation rather than indexing existing hot tables, so its exposure is much smaller — the same fix covers it. I read `scripts/migrate.ts` and it does not inject timeouts itself.
 - **Durable fix:** Move the guarantee into the runner instead of the file. Have `scripts/migrate.ts` apply `SET LOCAL lock_timeout` and `statement_timeout` around every migration it runs, so no future file can omit them, and add a check that refuses a production migration containing a non-concurrent `CREATE INDEX` unless it is explicitly acknowledged. Then rewrite this migration to use `CREATE INDEX CONCURRENTLY` (which cannot run inside a transaction and must be applied statement-by-statement with its own retry, so the runner needs to understand that mode).
 - **Why this is not a band-aid:** Adding the preamble to this one file fixes this migration; putting it in the runner makes the class of mistake impossible, which is what the two sibling files show was intended all along.
-- **Pre-fix proof:** Extend `test/deploy-safety.test.ts`, which already asserts properties of migration files (see "public pagination indexes are an explicitly selected additive release"). Add an assertion that every file in `db/migrations/` either sets both `lock_timeout` and `statement_timeout`, or uses `CREATE INDEX CONCURRENTLY` exclusively. Today it fails, naming `20260814_public_pagination.sql`.
+- **Pre-fix proof:** Extend `test/deploy-safety-tests/pagination-migration.test.ts`, which already asserts properties of migration files (see "public pagination indexes are an explicitly selected additive release"). Add an assertion that every file in `db/migrations/` either sets both `lock_timeout` and `statement_timeout`, or uses `CREATE INDEX CONCURRENTLY` exclusively. Today it fails, naming `20260814_public_pagination.sql`.
 - **Verification:** The test above; the full `test/migrate.test.ts` and `test/deploy-safety.test.ts`; and a rehearsal of the migration against a seeded scratch database with concurrent writers, measuring the write stall.
 - **Regression and rollback risk:** `CREATE INDEX CONCURRENTLY` can leave an invalid index behind if it fails, which then needs a manual drop — the runner must detect and report that rather than continuing. If this migration has already been applied to production, do not re-run it; fix the runner and the file for the next one and verify the existing indexes are valid.
 - **Unknowns:** Whether this migration has already been run against production, and the current row counts of `events` and `notes`, which set the real outage length. Both need production access.
